Move working code to new file.
